+++ lib/aoi.js/src/functions/discord/index.ts
@@ -2,4 +2,4 @@ export * from './client/index.js';
 export * from './guild/index.js';
 export * from './channel/index.js';
 export * from './message/index.js';
-export * from  './author/index.js';
+export * from  './user/index.js';
