+++ lib/aoi.js/src/functions/discord/user/$authoravatar.test.ts
