+++ lib/aoi.js/src/functions/discord/user/$authorbanner.test.ts
