+++ lib/aoi.js/src/functions/discord/user/$authorbanner.ts
@@ -9,15 +9,15 @@ import AoiError from '@aoi.js/core/Error.js';
  * @example
  * ```aoi
  * ---
- * name: authoravatar
+ * name: authorBanner
  * type: basic
  * ---
  *
- * $authoravatar // returns the banner of the author
+ * $authorBanner // returns the banner of the author
  * ```
  */
-const $authoravatar = new FunctionBuilder()
-	.setName('$authoravatar')
+const $authorbanner = new FunctionBuilder()
+	.setName('$authorbanner')
 	.setType(FunctionType.Getter)
 	.setReturns(ReturnType.String | ReturnType.Void)
 	.setBrackets(true)
@@ -43,22 +43,33 @@ const $authoravatar = new FunctionBuilder()
 		},
 	])
 	.setCode((data, scopes, thisArg) => {
+		const currentScope = thisArg.getCurrentScope(scopes);
 		const [size = '4096', dynamic = 'true', extension = 'png'] =
 			thisArg.getParams(data);
 
+		let res;
 		let parsedSize = thisArg.parseData(size, ReturnType.Number);
+		let parsedDyanmic = thisArg.parseData(dynamic, ReturnType.Boolean);
 
-		if (!thisArg.isCorrectType(parsedSize, ReturnType.Number)) {
+		if (!thisArg.isCorrectType(parsedSize, ReturnType.Number) && !thisArg.canSuppressAtComp(data, currentScope)) {
 			throw AoiError.FunctionError(
 				ErrorCode.InvalidArgumentType,
-				`Invalid type for parameter 'size' in function $authorAvatar, got ${parsedSize} expected: number.`,
+				`Invalid type for parameter 'size' in function $authorBanner, got ${parsedSize} expected: number.`,
 				data,
 			);
 		}
 
-		const result = thisArg.getResultString(
+		if (!thisArg.isCorrectType(parsedDyanmic, ReturnType.Boolean) && !thisArg.canSuppressAtComp(data, currentScope)) {
+			throw AoiError.FunctionError(
+				ErrorCode.InvalidArgumentType,
+				`Invalid type for parameter 'dynamic' in function $authorAvatar, got ${parsedDyanmic} expected: boolean.`,
+				data,
+			);
+		}
+
+		res = thisArg.getResultString(
 			(discordData) =>
-				discordData.author?.displayAvatarURL({
+				discordData.author?.bannerURL({
 					size: '$0',
 					// eslint-disable-next-line @typescript-eslint/ban-ts-comment
 					//@ts-ignore $1 is a placeholder
@@ -68,7 +79,8 @@ const $authoravatar = new FunctionBuilder()
 			[size, dynamic, parseString(extension)],
 		);
 
-		const escaped = escapeResult(result);
+
+		const escaped = escapeResult(res);
 
 		return {
 			code: escaped,
@@ -77,4 +89,4 @@ const $authoravatar = new FunctionBuilder()
 	})
 	.build();
 
-export { $authoravatar };
+export { $authorbanner };
