+++ lib/aoi.js/src/functions/discord/user/$authorbannercolor.test.ts
