+++ lib/aoi.js/src/functions/discord/user/$authorbannercolor.ts
