+++ lib/aoi.js/src/functions/discord/user/$finduser.test.ts
