+++ lib/aoi.js/src/functions/discord/user/$username.test.ts
