+++ lib/aoi.js/src/functions/discord/user/index.ts
